+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to
 
 ### Fixed
 
+- Drop webhook response body when status is 204 or 304
+  [#4778](https://github.com/OpenFn/lightning/issues/4778)
+
 ## [2.16.4-pre1] - 2026-05-19
 
 ### Fixed